sql-plugin compiles
