+++ sql-plugin-api/src/main/scala/com/nvidia/spark/rapids/ShimLoader.scala
@@ -24,8 +24,8 @@ import scala.util.Try
 
 import org.apache.commons.lang3.reflect.MethodUtils
 
-import org.apache.spark.api.plugin.{DriverPlugin, ExecutorPlugin}
 import org.apache.spark.{SPARK_BRANCH, SPARK_BUILD_DATE, SPARK_BUILD_USER, SPARK_REPO_URL, SPARK_REVISION, SPARK_VERSION, SparkConf, SparkEnv}
+import org.apache.spark.api.plugin.{DriverPlugin, ExecutorPlugin}
 import org.apache.spark.api.resource.ResourceDiscoveryPlugin
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.Strategy
